remove code duplication between shared_fs and metrics_storage
